Add endpoint client assignment management

+++ app/routes.php
@@ -75,6 +75,11 @@ if (preg_match('#^/endpoints/(\d+)/toggle$#', $path, $matches) && $method === 'P
     exit;
 }
 
+if (preg_match('#^/endpoints/(\d+)/clients$#', $path, $matches) && $method === 'POST') {
+    \PrintBridge\Controllers\EndpointController::assignClients((int) $matches[1]);
+    exit;
+}
+
 if (preg_match('#^/endpoints/(\d+)/delete$#', $path, $matches) && $method === 'POST') {
     \PrintBridge\Controllers\EndpointController::delete((int) $matches[1]);
     exit;
